refactor: checkStopsWithinTripShape() into StopTooFarFromTripShapeValidator
