+++ main/src/main/java/org/mobilitydata/gtfsvalidator/validator/StopTooFarFromTripShapeValidator.java
@@ -16,16 +16,25 @@
 
 package org.mobilitydata.gtfsvalidator.validator;
 
+import static org.mobilitydata.gtfsvalidator.util.GeospatialUtil.TRIP_BUFFER_DEGREES;
+import static org.mobilitydata.gtfsvalidator.util.GeospatialUtil.TRIP_BUFFER_METERS;
+
 import com.google.common.collect.Multimaps;
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map.Entry;
 import java.util.Set;
+import org.locationtech.spatial4j.shape.Point;
+import org.locationtech.spatial4j.shape.Shape;
+import org.locationtech.spatial4j.shape.ShapeFactory;
+import org.locationtech.spatial4j.shape.SpatialRelation;
 import org.mobilitydata.gtfsvalidator.annotation.GtfsValidator;
 import org.mobilitydata.gtfsvalidator.annotation.Inject;
 import org.mobilitydata.gtfsvalidator.notice.NoticeContainer;
 import org.mobilitydata.gtfsvalidator.notice.StopTooFarFromTripShapeNotice;
+import org.mobilitydata.gtfsvalidator.table.GtfsLocationType;
+import org.mobilitydata.gtfsvalidator.table.GtfsShape;
 import org.mobilitydata.gtfsvalidator.table.GtfsShapeTableContainer;
 import org.mobilitydata.gtfsvalidator.table.GtfsStop;
 import org.mobilitydata.gtfsvalidator.table.GtfsStopTableContainer;
@@ -67,7 +76,7 @@ public class StopTooFarFromTripShapeValidator extends FileValidator {
       // Check for possible errors for this combination of stop times and shape points for
       // this trip_id
       List<StopTooFarFromTripShapeNotice> noticesForTrip =
-          GeospatialUtil.checkStopsWithinTripShape(
+          checkStopsWithinTripShape(
               tripId,
               stopTimesForTrip,
               trip.shapeId(),
@@ -79,5 +88,73 @@ public class StopTooFarFromTripShapeValidator extends FileValidator {
     for (StopTooFarFromTripShapeNotice notice : notices) {
       noticeContainer.addValidationNotice(notice);
     }
+  }
+
+  /**
+   * Returns a list of notices for the given input, one for each stop that is too far from the trip
+   * shapePoints
+   *
+   * @param tripId trip_id for this GTFS trip
+   * @param stopTimes a list of StopTimes for a trip, sorted by stop_sequence
+   * @param shapeId the shape_id for this GTFS trip
+   * @param shapePoints a list of ShapePoints for a trip, sorted by shape_pt_sequence
+   * @param stopTable a container for all stops (keyed on stop_id), needed to obtain the latitude
+   *     and longitude for each stop on the trip
+   * @param testedCache a cache for previously tested shape_id and stop_id pairs (keyed on
+   *     shape_id+stop_id). If the combination of shape_id and stop_id appears in this set, we
+   *     shouldn't test it again. Shapes and stops tested in this method execution will be added to
+   *     this testedCache.
+   * @return a list of notices, one for each stop that is too far from the trip shape
+   */
+  private List<StopTooFarFromTripShapeNotice> checkStopsWithinTripShape(
+      final String tripId,
+      final List<GtfsStopTime> stopTimes,
+      final String shapeId,
+      final List<GtfsShape> shapePoints,
+      final GtfsStopTableContainer stopTable,
+      final Set<String> testedCache) {
+    List<StopTooFarFromTripShapeNotice> notices = new ArrayList<>();
+    if (stopTimes == null || stopTimes.isEmpty() || shapePoints == null || shapePoints.isEmpty()) {
+      // Nothing to do - return empty list
+      return notices;
+    }
+
+    // Create a polyline from the GTFS shapes data
+    ShapeFactory.LineStringBuilder lineBuilder = GeospatialUtil.getShapeFactory().lineString();
+    for (GtfsShape shapePoint : shapePoints) {
+      lineBuilder.pointXY(shapePoint.shapePtLon(), shapePoint.shapePtLat());
+    }
+    Shape shapeLine = lineBuilder.build();
+
+    // Create the buffered version of the trip as a polygon
+    Shape shapeBuffer = shapeLine.getBuffered(TRIP_BUFFER_DEGREES, shapeLine.getContext());
+
+    // Check if each stop is within the buffer polygon
+    for (GtfsStopTime stopTime : stopTimes) {
+      GtfsStop stop = stopTable.byStopId(stopTime.stopId());
+      if (stop == null || !stop.hasStopLat() || !stop.hasStopLon()) {
+        // Lat/lon are optional for location_type 4 - skip to the next stop if they aren't
+        // provided
+        continue;
+      }
+      if (!(stop.locationType() == GtfsLocationType.STOP)
+          && !(stop.locationType() == GtfsLocationType.BOARDING_AREA)) {
+        // This rule only applies to stops of location_type 0 and 4 - skip to next stop
+        continue;
+      }
+      if (testedCache.contains(shapeId + stop.stopId())) {
+        // We've already tested this combination of shape ID and stop ID - skip to next stop
+        continue;
+      }
+      testedCache.add(shapeId + stop.stopId());
+      Point p = GeospatialUtil.getShapeFactory().pointXY(stop.stopLon(), stop.stopLat());
+      if (!shapeBuffer.relate(p).equals(SpatialRelation.CONTAINS)) {
+        notices.add(
+            new StopTooFarFromTripShapeNotice(
+                stopTime.stopId(), stopTime.stopSequence(), tripId, shapeId, TRIP_BUFFER_METERS));
+      }
+    }
+
+    return notices;
   }
 }
